#536 added tests for get/assessments router
